+++ package.json
@@ -41,7 +41,6 @@
     "@typescript-eslint/eslint-plugin": "^5.15.0",
     "@typescript-eslint/parser": "^5.15.0",
     "axios": "^0.26.1",
-    "dayjs": "^1.11.0",
     "discord.js": "^13.6.0",
     "dotenv": "^16.0.0",
     "eslint": "^8.11.0",
@@ -50,7 +49,6 @@
     "node-fetch": "^3.2.3",
     "prettier": "^2.6.0",
     "prettier-plugin-svelte": "^2.6.0",
-    "query-string": "^7.1.1",
     "sass": "^1.49.9",
     "size-limit": "^7.0.8",
     "svelte": "^3.46.4",
