--- src/lib/calendar-link/README.md
@@ -1,6 +1,0 @@
-# `calendar-link`
-
-## Credit
-
-- Demo: <https://anandchowdhary.github.io/calendar-link/>
-- Repository: <https://github.com/AnandChowdhary/calendar-link>
