--- src/lib/calendar-link/index.ts
@@ -1,1 +1,0 @@
-export * from './calendar-link';
