--- src/lib/calendar-link/utils.ts
@@ -1,5 +1,0 @@
-export const TimeFormats = {
-  dateTimeWithOffset: 'YYYY-MM-DD[T]HH:mm:ssZ',
-  dateTimeUTC: 'YYYYMMDD[T]HHmmss[Z]',
-  allDay: 'YYYYMMDD',
-};
